+++ krokiet/src/common.rs
@@ -1,8 +1,8 @@
 #![allow(dead_code)] // TODO later remove
 use std::path::PathBuf;
 
-use crate::{CurrentTab, ExcludedDirectoriesModel, IncludedDirectoriesModel, MainListModel, MainWindow};
-use slint::{ModelRc, SharedString, VecModel};
+use crate::{CurrentTab, ExcludedDirectoriesModel, IncludedDirectoriesModel, MainListModel, MainWindow, Settings};
+use slint::{ComponentHandle, Model, ModelRc, SharedString, VecModel};
 
 // Int model is used to store data in unchanged(* except that we need to split u64 into two i32) form and is used to sort/select data
 // Str model is used to display data in gui
@@ -371,6 +371,11 @@ pub fn create_excluded_directories_model_from_pathbuf(items: &[PathBuf]) -> Mode
         })
         .collect::<Vec<_>>();
     ModelRc::new(VecModel::from(converted))
+}
+
+pub fn check_if_all_included_dirs_are_referenced(app: &MainWindow) -> bool {
+    let included = app.global::<Settings>().get_included_directories_model();
+    included.iter().all(|x| x.referenced_folder)
 }
 
 pub fn create_vec_model_from_vec_string(items: Vec<String>) -> VecModel<SharedString> {